adding template changing example (#109)
* Add files via upload

* Delete theme_changer.py

* Add files via upload

Allows user to select which plotly template to use on the graphs.

* Add files via upload

I added this countdown app which is very similar to the stopwatch app. This one has some tweaks such as controlling the time format (not just seconds but e.g. hours, minutes and seconds if given large enough integer), using chained callbacks and audio. The audio file is in the assets folder of my local directory so I was not sure how to add it. The file is available at https://cdn.pixabay.com/download/audio/2021/09/27/audio_91211934db.mp3?filename=clock-alarm-8761.mp3

* Update examples/template-change.py

* Add files via upload

I updated the IDs to the format of "template-change-x-"

* Add files via upload

I updated the ids and removed the brackets

* Add files via upload

* Add files via upload

I changed the asset folder path to relative path and added the audio file to the assets folder

Co-authored-by: Adam <adam@charming-data.com>