Update examples/timeline-gantt.py
It's challenging to read the callback when this function is outside of it. I had to refer to this function a few times (scrolling back and forth) before fully understanding what was going on. I think it is much easier to read if we put this function inside the callback function, since there is a lot going on here that is directly tied to the callback decorator and function.