Update key-figures-finland (#128)
* Add files via upload

* Delete theme_changer.py

* Add files via upload

Allows user to select which plotly template to use on the graphs.

* Add files via upload

I added this countdown app which is very similar to the stopwatch app. This one has some tweaks such as controlling the time format (not just seconds but e.g. hours, minutes and seconds if given large enough integer), using chained callbacks and audio. The audio file is in the assets folder of my local directory so I was not sure how to add it. The file is available at https://cdn.pixabay.com/download/audio/2021/09/27/audio_91211934db.mp3?filename=clock-alarm-8761.mp3

* Update examples/template-change.py

* Add files via upload

I updated the IDs to the format of "template-change-x-"

* Add files via upload

I updated the ids and removed the brackets

* Add files via upload

* Add files via upload

I changed the asset folder path to relative path and added the audio file to the assets folder

* Create keyfiguresfinland.py

* Added files for the keyfiguresfinland.py

* Update keyfiguresfinland.py

I changed the whole generation of the map to the clientside. Now the app works really fast.

* Update keyfiguresfinland.py

* Delete countdown.py

* Delete template-change.py

* Delete clock-alarm-8761.mp3

* Add files via upload

* Delete assets_clock-alarm-8761.mp3

* Create keyfiguresfinland_serverside.py

Here is the same Finland's key figures app but with a serverside callback on updating the map.

* Added Code Review checklist

* Update keyfiguresfinland.py

I made some changes: 
- Moved the extraction of header fields to the beginning.
- Simplified the header update callback.
- Added 'white-space': 'nowrap' to the dropdown style to prevent overlapping on mobile

* Update keyfiguresfinland.py

* Add files via upload

Added new json files since Åland isles did not render on the map for some reason.

* Update keyfiguresfinland.py

Went through the code to make ends meet better on the checklist.
Basic requirements:
  - replaced orjson with json since orjson is not on the requirements list
  - removed app.title and server = app.server
  - minimized styling (removed font size etc.)
  - has no errors in the console
Naming conventions:
   - uses df in the global scope and dff in callbacks
Concise code:
   - removed [{'label':c, 'value':c} for c in df.columns] notation from dcc.Dropdown
   - removed multi = False from the dropdown
   - tried to add 'why' in the comments
Bootstrap:
   - replaced styles with classnames
   - replaced html.Br() with classnames
   - added dbc.themes.SPACELAB for testing, works well

* Update keyfiguresfinland.py

Made changes to styling and removed header from above the graph.

* Update keyfiguresfinland.py

corrected zoom also on graph initialization if it makes any difference :)

* Delete keyfiguresfinland_serverside.py

* Delete keyfiguresfinland.py

* Add files via upload

* Add files via upload

* Delete key-figures-finland.py

* Delete CONTRIBUTING.md

* Update key-figures-finland.py

* Update key-figures-finland.py

It's formatted with black

* Add files via upload

* read json file from plotly data sets repo

* read csv file from plotly data sets repo

* Update examples/key-figures-finland.py

* add requests library

* added utf-8 enconding to csv

* delete space

* format numbers with comma instead of space

* remove the empty line accidently added

* Delete key_figures_regions.csv

* Delete key-figures-finland.py

* Delete regions.json

* Add files via upload

* Update key-figures-finland.py

* Update key-figures-finland.py

Co-authored-by: Adam <adam@charming-data.com>
Co-authored-by: AnnMarieW <amward@fastmail.com>